requirements.txt
python-dotenv
discord.py
aioredis
asyncio
pyqrcode
pypng
requests
paho-mqtt